Refactored into core/cloud feature stores
